Add files via upload
Day 1:
We planned lot of things and brought everything in track in first few days. Moreover the only reason of you being in this exclusive team is that you will be organized by our help and will be consistent. Similarly, if you face any issue, then you can post that in the classroom or if it is specific, don't hesitate to text me.

In the beginning we will be start with slow pace. Those of you who have been actively solving problem with us from last 15 days, you might be better than those who have joined us now. But in the matter of the week, we will all start learning in the same pace.

So the 8 a.m problem of today are bit simple to solve to boost your confidence. 

Create:  https://help.github.com/en/desktop/getting-started-with-github-desktop
Problem 1: Implement Linkedlist without looking out for any help. Don't google if you already done that. Insertion, Deletion and searching element recursively.
Problem 2: Have you heard of loop in the linkedlist? If yes, how do we find it? Write a code in language of your choice.

Push the solution of this code on your github repository and comment the link to your github repository in the comments. I'll look into your codes and will suggest you flaws so that you can improve on top of that.