Create LongestSubstring.cpp
Day 19:
Problem 1:
Given a string, find the length of the longest sub-string without repeating characters. For example, the longest sub-strings without repeating characters for “ABDEFGABEF” are “BDEFGA” and “DEFGAB”, with length 6.

Example 1:

Input: "abcabcbb"
Output: 3 
Explanation: The answer is "abc", with the length of 3. 
Example 2:

Input: "bbbbb"
Output: 1
Explanation: The answer is "b", with the length of 1.
Example 3:

Input: "pwwkew"
Output: 3
Explanation: The answer is "wke", with the length of 3. 
             Note that the answer must be a substring, "pwke" is a subsequence and not a substring.

This problem is quite popular but solving it with brute force can be easy, but try with sliding window approach to reduce the complexity.