Create Min_xorValue.c
Day 4:
Problem solving

If the yesterday was not tough, let's make today a learning experience. So just try these questions.
Problem 1: Reverse the bits of an 32 bit unsigned integer A.
Problem 2: Given an integer array A of N integers, find the pair of integers in the array which have minimum XOR value. Report the minimum XOR value.

Just share your queries in the comment and we shall discuss the importance and how these questions can be leveraged!!

But you might be having a question, when will we do the project?
Very soon. 
" Slow and steady wins the race", Let's be steady to see the results!!