Create DistinctSubsets.cpp
Day 16:

I hope you all were able to do the questions. Many of you faced lot of issues. But that's okay.
So the problem for today is same but have different approach to solve them.

Problem 1: Find out all the possible subsets of a given set using recursion. We did this question in bit manipulation.

Problem 2: Now that you have solved this, what about not having duplicated subsets in our answer when the given set which is repesented as array contains duplicate element. ( Use recursion in both the problems)