day_24
Day 24:
Almost all of you might have done at least 80% of what has been discussed. Well, there are still few things left for discussion which I would love to do over weekends when I find plenty of time. But anyway, I have been open to your personal texts 24*7 and has been stressing from my side to know what's your problem by holding even personal session which is quite lengthy process.
Let' solve a very traditional problem for today. You can do it in an hour but I'm sure you have never thought on it's application which most of people don't do.

This problems is Game of Death in a circle.
If you would like to know in detail just like Debnath did for Euclid's algorithm, you must spend your time reading https://mathworld.wolfram.com/JosephusProblem.html

P.S. It is lengthy.

Problem :


There are n people standing in a circle (numbered clockwise 1 to n) waiting to be executed. The counting begins at point 1 in the circle and proceeds around the circle in a fixed direction (clockwise). In each step, a certain number of people are skipped and the next person is executed. The elimination proceeds around the circle (which is becoming smaller and smaller as the executed people are removed), until only the last person remains, who is given freedom.
Given the total number of persons n and a number k which indicates that k-1 persons are skipped and kth person is killed in circle. The task is to choose the place in the initial circle so that you are the last one remaining and so survive.
Consider if n = 5 and k = 2, then the safe position is 3.
Firstly, the person at position 2 is killed, then person at position 4 is killed, then person at position 1 is killed. Finally, the person at position 5 is killed. So the person at position 3 survives.

So, just print the safe positions.